+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/actions/list/cds-action-askgpt-v2/cds-action-askgpt-v2.component.ts
@@ -19,6 +19,7 @@ import { TYPE_UPDATE_ACTION, TYPE_GPT_MODEL } from 'src/app/chatbot-design-studi
 import { variableList } from 'src/app/chatbot-design-studio/utils-variables';
 import { Namespace } from 'src/app/models/namespace-model';
 import { TranslateService } from '@ngx-translate/core';
+import { loadTokenMultiplier } from 'src/app/utils/util';
 
 @Component({
   selector: 'cds-action-askgpt-v2',
@@ -55,7 +56,7 @@ export class CdsActionAskgptV2Component implements OnInit {
 
   temp_variables = [];
 
-  model_list: Array<{ name: string, value: string }>;
+  model_list: Array<{ name: string, value: string, multiplier: string}>;
   //namespace_list = []; // missing api for namespaces
 
   private subscriptionChangedConnector: Subscription;
@@ -72,7 +73,13 @@ export class CdsActionAskgptV2Component implements OnInit {
 
   ngOnInit(): void {
     this.logger.debug("[ACTION-ASKGPTV2] action detail: ", this.action);
-    this.model_list = Object.values(TYPE_GPT_MODEL).filter(el=> el.status !== 'inactive')
+    const ai_models = loadTokenMultiplier(this.appConfigService.getConfig().aiModels)
+    this.model_list = Object.values(TYPE_GPT_MODEL).filter(el=> el.status !== 'inactive').map((el)=> {
+      if(ai_models[el.value])
+        return { ...el, multiplier: ai_models[el.value] + ' x tokens' }
+      else
+        return { ...el, multiplier: null }
+    })
     if (!this.action.namespace) {
       this.action.namespace = this.project_id;
     }
@@ -332,11 +339,12 @@ export class CdsActionAskgptV2Component implements OnInit {
 
     let data = {
       question: question,
-      //context: this.action.context,
+      context: this.action.context,
       model: this.action.model,
       max_tokens: this.action.max_tokens,
       temperature: this.action.temperature,
-      top_k: this.action.top_k
+      top_k: this.action.top_k,
+      namespace: this.action.namespace
     }
 
     this.showAiError = false;