+++ src/app/chatbot-design-studio/utils-errors.ts
@@ -1,0 +1,4 @@
+export const AiErrors: { [key: string] : string} = {
+    "13001": "CDSCanvas.Errors.QuoteExceeded",
+    "13002": "CDSCanvas.Error.General"
+}